# perf: avoid allocations with `(*regexp.Regexp).MatchString`