@nitrocode 2023-10-21
Fwiw I agree with this change.

I see some references online that also point to a performance gain

> They are very similar. while slices like []byte are mutable (you can change values in the underlying array) strings are immutable.
Internally []byte consists of a length, capacity (both int values) and a pointer to the first element [1]. Strings only consist of the length and the pointer because capacity and length are always the same since strings are immutable.
> > What are the dangers of casting between them?
>
> Wasted performance and memory
>
> >Why are regexp functions even relevant for []byte?
>
> Because it is cheap. The function can easily return a sub-slice without copying any data.

https://groups.google.com/g/golang-nuts/c/YmiYPe7lako?pli=1