@jamengual 2023-11-15
/cherry-pick release-0.26